+++ backend/controllers/scoreControllers.js
@@ -1,20 +1,5 @@
 import PlayerScore from "../models/scores.js";
 
-// Serve the home page
-export const getHomePage = (req, res) => {
-  res.json({ message: 'Welcome to Tetris! Click "Play" to start.' });
-};
-
-// Serve the game page
-export const getGamePage = (req, res) => {
-  res.json({ message: "This is where the Tetris game will load." });
-};
-
-// Serve the leaderboard page
-export const getLeaderboardPage = (req, res) => {
-  res.json({ message: "Leaderboard will be displayed here." });
-};
-
 // Handle score submissions
 export const submitNewScore = async (req, res) => {
   const { player, score, date } = req.body;  // Extract player, score, and date from the request body
