+++ backend/index.js
@@ -1,7 +1,8 @@
-import express from "express";
-import mongoose from "mongoose";
-import dotenv from "dotenv"; // Use `import` for dotenv
-import navigationRoutes from './routes/navigation.js'; // Import with `.js` extension
+import express from 'express';
+import mongoose from 'mongoose';
+import dotenv from 'dotenv';
+import navigationRoutes from './routes/navigation.js'; // Navigation routes
+import scoreRoutes from './routes/scores.js';           // Score routes
 import cors from 'cors';
 
 // Load environment variables
@@ -12,7 +13,6 @@ const app = express();
 
 // Middleware
 app.use(express.json());
-
 app.use(cors());
 
 app.use((req, res, next) => {
@@ -20,15 +20,14 @@ app.use((req, res, next) => {
     next();
 });
 
-// Use the navigation routes
-app.use("/api/navigation", navigationRoutes);
+// Use routes
+app.use('/navigation', navigationRoutes);  // Navigation-related routes
+app.use('/api', scoreRoutes);                  // Score-related routes
 
 // Connect to the database
-mongoose
-    .connect(process.env.DB_CONNECTION_STRING)
+mongoose.connect(process.env.DB_CONNECTION_STRING)
     .then(() => {
         console.log("App connected to database");
-
         app.listen(process.env.PORT, () => {
             console.log(`App is listening on port: ${process.env.PORT}`);
         });
