+++ backend/routes/navigation.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { getHomePage, getGamePage, getLeaderboardPage, submitNewScore, getLeaderboardData } from '../controllers/scoreControllers.js';
+import { getHomePage, getGamePage, getLeaderboardPage} from '../controllers/navigationControllers.js';
 
 const router = Router();
 
@@ -11,11 +11,5 @@ router.get('/play', getGamePage);
 
 // Serve the leaderboard page
 router.get('/leaderboard', getLeaderboardPage);
-
-// Handle score submissions (API)
-router.post('/newScore', submitNewScore);
-
-// Retrieve leaderboard data (API)
-router.get('/api/scores', getLeaderboardData);
 
 export default router;
