+++ frontend/src/pages/Leaderboard.jsx
@@ -8,7 +8,7 @@ const Leaderboard = () => {
   useEffect(() => {
     const fetchLeaderboardData = async () => {
       try {
-        const response = await fetch('http://localhost:3000/api/navigation/api/scores');
+        const response = await fetch('http://localhost:3000/api/scores');
         if (!response.ok) {
           throw new Error('Failed to fetch leaderboard data');
         }
